data-analysis/requirements.txt
# Python依赖包清单
# 安装命令: pip install -r requirements.txt

# 数据处理
numpy>=1.21.0
pandas>=1.3.0

# 数据可视化
matplotlib>=3.4.0

# 数据格式支持
pyyaml>=5.4.0

# 可选：更好的命令行输出
colorama>=0.4.4

# 可选：进度条
tqdm>=4.62.0

